refac transfer inject test
